auth

+++ controllers/contactsControllers.js
@@ -1,10 +1,29 @@
 import contactsService from "../services/contactsServices.js";
 import HttpError from "../helpers/HttpError.js";
+import ctrlWrapper from "../decorators/ctrlWrapper.js";
 
 export const getAllContacts = async (req, res, next) => {
     try {
-        const contacts = await contactsService.listContacts();
-        res.status(200).json(contacts);
+        const { id: owner } = req.user;
+        //pagination
+        const page = parseInt(req.query.page) || 1;
+        const limit = parseInt(req.query.limit) || 10;
+        const offset = (page - 1) * limit;
+
+        //filter favorites
+        const filter = { owner };
+        if (req.query.favorite !== undefined) {
+            filter.favorite = req.query.favorite === "true";
+        }
+
+        const { count, rows: contacts } = await contactsService.listContacts(filter, limit, offset);
+
+        res.status(200).json({
+            total: count,
+            page,
+            limit,
+            contacts,
+        });
     } catch (error) {
         next(error);
     }
@@ -13,7 +32,8 @@ export const getAllContacts = async (req, res, next) => {
 export const getOneContact = async (req, res, next) => {
     try {
         const { id } = req.params;
-        const contact = await contactsService.getContactById(id);
+        const { id: owner } = req.user;
+        const contact = await contactsService.getContact({ id, owner });
         if (!contact) {
             throw HttpError(404, "Not found");
         }
@@ -26,13 +46,12 @@ export const getOneContact = async (req, res, next) => {
 export const deleteContact = async (req, res, next) => {
     try {
         const { id } = req.params;
-        const removedContact = await contactsService.removeContact(id);
+        const { id: owner } = req.user;
+        const removedContact = await contactsService.removeContact({ id, owner });
         if (!removedContact) {
             throw HttpError(404, "Not found");
         }
-        res.json({
-            message: "Delete successfully",
-        });
+        res.status(200).json(removedContact);
     } catch (error) {
         next(error);
     }
@@ -40,7 +59,8 @@ export const deleteContact = async (req, res, next) => {
 
 export const createContact = async (req, res, next) => {
     try {
-        const newContact = await contactsService.addContact(req.body);
+        const { id: owner } = req.user;
+        const newContact = await contactsService.addContact({ ...req.body, owner });
         res.status(201).json(newContact);
     } catch (error) {
         next(error);
@@ -50,7 +70,8 @@ export const createContact = async (req, res, next) => {
 export const updateContact = async (req, res, next) => {
     try {
         const { id } = req.params;
-        const updatedContact = await contactsService.updateContact(id, req.body);
+        const { id: owner } = req.user;
+        const updatedContact = await contactsService.updateContact({ id, owner }, req.body);
         if (!updatedContact) {
             throw HttpError(404, "Not found");
         }
@@ -62,10 +83,11 @@ export const updateContact = async (req, res, next) => {
 
 export const updateStatusContact = async (req, res, next) => {
     try {
-        const { contactId } = req.params;
+        const { id } = req.params;
         const { favorite } = req.body;
+        const { id: owner } = req.user;
 
-        const updatedContact = await contactsService.updateStatusContact(contactId, { favorite });
+        const updatedContact = await contactsService.updateStatusContact({ id, owner }, { favorite });
 
         if (!updatedContact) {
             throw HttpError(404, "Not found");
@@ -75,4 +97,13 @@ export const updateStatusContact = async (req, res, next) => {
     } catch (error) {
         next(error);
     }
+};
+
+export default {
+    getAll: ctrlWrapper(getAllContacts),
+    getById: ctrlWrapper(getOneContact),
+    add: ctrlWrapper(createContact),
+    updateById: ctrlWrapper(updateContact),
+    deleteById: ctrlWrapper(deleteContact),
+    updateStatus: ctrlWrapper(updateStatusContact),
 };
